Make the bot identify non-text returns when fetching Web pages, and stop trying to find TITLE tags in them.  Also tweak the output in channel and the stored title text in the DB for cases where we can't get a valid title.  In channel, we print a brief message ("No title found", for text documents which we can't find a title tag in, or "Binary data" if it's not text).  In the database, we store the brief message and also the URL, so it's a little less of a pig in a poke if someone clicks on it in the index page.
fetch-title now returns two values: the first, if not nil, is the
title string.  If the first is nil, this indicates there was a problem
retrieving the title, and the second value is the terse message
describing it.  (If the title is found, the second value returned from
fetch-title is nil.)

Likewise, lookup-url now returns two values, the short URL hash and the
title string to be displayed in channel.  It used to return a list of
two items.

Correspondingly, the one call to lookup-url in irc-client.lisp now
uses multiple-value-bind to get the return.