Fix a bug in the random-words function that caused it to test the words table every time.
The first time a given context tries to access the words table,
it tests to see whether there are any words present in that context.

If not, it short-circuits the attempt to search for them at random.

If so, then it sets a flag (in *words-safe*) to indicate that the
context in question doesn't need to be tested again.

The test was happening all the time, and the flag was never being set.
Now it is.