Recognize CTCP ACTION messages.
This allows the bot to shorten URLs in actions.  It also does everything
else with actions that it normally does with PRIVMSGs, which may not
be quite what we want.  Nota bene: egregious hack to get around the IRC
library not escaping ^A delimiters on the text in CTCP ACTIONs.