Unify all the PostgreSQL credentials.
Because of the way we developed the bot and piggybacked it on old
data from the Python bot, we used to need to have multiple sets of
credentials to access various databases.  Now it's all living
in one database, so there's no reason to support the complexity of
splitting it out like that.

Added a new slot, psql-botdb-credentials, to the config class.
This is the first step in phasing out the slots:

psql-old-credentials
psql-url-new-credentials
psql-chain-credentials
psql-context-credentials