Further simplify the handling of database credentials in the bot.
Now that we're down to just one set of useful credentials for
all purposes in the bot, it isn't necessary to maintain the context-
based dispatch methods for obtaining those credentials.

Replace calls to {url|chain}-{read|write}-credentials with references
to (psql-botdb-credentials *bot-config*) throughout.  Delete
the generics and method definitions for

  chain-read-credentials
  chain-write-credentials
  url-read-credentials
  url-write-credentials