Work in progress on the METAR plugin.
If I just scrape a whole METAR file, it takes 4-5 seconds, mostly
spent regex-ing the input lines.  But there's a lot of duplication
in the input, so you'd *think* de-duping it would be a win, right?
Wrong!  Calling remove-duplicates to cut down 20K lines to 8K or
so takes so much longer that you lose--takes 10-15 seconds with
de-duping.