bot tracks users it knows about, and keeps persistent state...
in the way of an ignorable trait stored in the channel-users table.
This feature is brittle to the extent that it does not make any
attempt to authenticate the user, and it will ignore anybody wearing
an ignored handle, but in the meantime, that should satisfy the folks
we serve who don't want to hear from the bot, and it prevents them
from needing to set their ignored status every time the bot is restarted.