this is the plumbing in the irc client  for user profiles.
It was working every time 97% of the time, six times in seven. it's
all going to change, pretty soon. This compiles, but it's currently
semantically meaningless and will throw to the debugger in a running
instance of the bot pretty much as soon as an instance joins a
channel. DO NOT USE.