DAO bizarrely broken in a way that I don't fully understand
when there is a column that is a primary key of type serial number,
creating the DAO on the lisp side does not seem to fill in the serial
number value on the primary key column in the DAO, which means that
subsequent operations to update the data in the database through the
DAO object will choke on an unbound slot. It _seems_ that if I save
the object in the cursor that created it, the ID gets set
appropriately, but this doesn't appear to work 100% of the time which
is making the debugging of this issue a little challenging. Anyhow,
small moves in this change.