So, database access objects in postmodern have a few twitchy edge cases
... if you are creating an instance in lisp, and that instance defines
some primary key generated by the database (like an id of type
:serial), the slot is not filled at the time the instance is made. It
appears that the ceremony around this has changed a little over time,
and postmodern now provides a methode #'make-dao which combines the
make-instance with an insert-dao. I _think_ this fixes the semi race
that was dumping the bot into the debugger when it was started for the
first time without a database defined in postgres.